Compact relocations
Instead of simply assuming every relocaiton is 32 bits, we take an
iterative approach to writing out the relocations.  We try assuming
everything in an object fits in 6 bits then move on to 8 then 16 and
finally we write 32-bit offsets if we absolutely have to.  This still
isn't quite as compact as the code produced by LabView in some cases but
it should be pretty good.